#1550: ddress PR #1727 review: KDoc, narrowed catch, required fallback for native output, BOM version management, logging, Java interop test, package move to api/tool/hitl
