+++ embabel-agent-api/pom.xml
@@ -129,7 +129,6 @@
         <dependency>
             <groupId>com.github.victools</groupId>
             <artifactId>jsonschema-module-jakarta-validation</artifactId>
-            <version>4.38.0</version>
         </dependency>
 
         <!-- SpotBugs annotations for thread-safety documentation (@ThreadSafe, @GuardedBy) -->
